♻️ Extract cache and helpers from AOJ API client (#1995)
